BonsaiBuddy_web/requirements.txt
bcrypt==3.2.2
django==4.2
pymongo==4.3.3
pycountry==22.3.5
mongoengine==0.27.0
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.0.0
django-tagify2==1.0.19
django-widget-tweaks==1.4.12
gunicorn==21.2.0
whitenoise==6.5.0
markdown==3.4.4
djangorestframework==3.14.0
django-rest-framework-mongoengine==3.4.1